backend/src/gateway/requirements.txt
altgraph==0.17.4
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-PyMongo==3.0.1
future==1.0.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
macholib==1.16.3
MarkupSafe==3.0.2
pika==1.3.2
PyJWT==2.10.1
pymongo==4.11.1
python-dotenv==1.0.1
six==1.17.0
Werkzeug==3.1.3
zipp==3.21.0
